pyproject.toml
[tool.poetry]
name = "pdrdeeplearning"
version = "0.1.0"
description = ""
authors = ["pdrsnow <229269179@qq.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"

pandas = "^2.2.2"
matplotlib = "^3.8.4"
scikit-learn = "^1.5.0"
pyyaml = "^6.0.1"
psycopg2-binary = "^2.9.9"
mysql-connector-python = "^8.4.0"
# 使用 cpu
torch = { version = "^2.3.0", source = "tsinghua" }
tensorflow-intel = { version = "^2.16.1", source = "tsinghua" }
# 使用 gpu, cuda <=11.8, cuDNN <= 8.9
# torch = { version = "^2.3.0+cu118", source = "torch" }
# tensorflow-gpu = { version = "<2.11", source = "aliyun" }
tensorflow-io-gcs-filesystem = { version = "^0.31.0", "python" = ">=3.10,<3.12" }
fastapi = "^0.111.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# 清华的镜像有问题, 某些包无法下载
[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"


[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu118/"
priority = "supplemental"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple"